Declares PHP version constraints
The code depends explicitly on PHP 5.4 [short array notation][1],
declaring it makes for a better conflict resolution for applications
depending on this package.

The core functionality relies on [GD][] functions, declaring it also
enables developers depending on the package to properly address this
dependency without having to discover it through an "function does not
exists" error. I tried to constraint `ext-gd` to a sensible semantic
versioning strategy but as a lot of functions were already available
under PHP 4 I saw as sensible enough allowing a wild card as version.
Some constraints should be made more explicitly depending on the type of
image being handled - which should be solved through the exposition of
the resources with their declaration under the `composer.json` file.

A minor correction on a comma (,) was made to conform the JSON output
generated by the `composer require` command - and the JSON
specification.

[1]: http://php.net/manual/en/language.types.array.php
[GD]: http://php.net/manual/en/book.image.php