Aids expectation over automated test execution
By providing a `phpunit.xml.dist` in place of `phpunit.xml` allows
developers to customize their preferred configuration along with
providing a default configuration which is automatically used by
PHPUnit. In short terms: it provides the same functionality
`phpunit.xml` does allowing developers to input their preference over
the defaults as well - a win win situation.

The test suite is configured to use composer autoload and provides
default configuration for PHP allowing the execution of the test suite
successfully without depending on the configuration of the host. The
skeleton was copied from `vendor/phpunit/phpunit/phpunit.xml`.

The fixture being used by the tests was not being found because of the
existence of the "." in its file name and the change of the test
directory from "./tests" to "./tests". PHP functions (and Unix) defaults
to not adding the last slash ("/") into paths, the change to the test
directory on `phpunit.xml.dist` was made to conform that - prompting the
changes into the test paths as well.

When errors occur on tests suites, one expects them to instruct
developers on how to fix (or easily expose the problem, easing debug),
some changes were made into the test suite and the Exception on the
concrete class allowing the error to explicitly tell what was not found -
allowing a developer to actually see the path and detect the problem to
be solve.

Exposes a `test` composer script to allow execution of the test suite
using `composer test` command. Although this is not a formal norm, it is
common to use that through out other languages as well. The "retab" of
the file was made automatically by the editor - I kept it since the file
had both spaces and tabs, it is consistent now at least.