Conforms to PSR2 standards
Standards are there to be used, [developers perform better with code
they are familiar][1] so conforming to expected standards is important.

Along with adding a tool to fix the code, provides a `composer ci`
script which tests for code standard violations as the fixer does not
fix all problems - using more than 120 chars on a single line, being an
example.

Some manual changes were made to conform PSR2:

* `SetUp` method conforming to camel-case (`setUp`).
* 120-length char line violations.
* Tests without a namespace declaration.

[1]: https://static.googleusercontent.com/media/research.google.com/en//pubs/archive/43835.pdf