Describes colors as classes
Better than providing a JSON, classes allow for so many things! They can
conform to an interface which is hidden in the JSON declaration
previously used - the public interface allow for the creation of new
colors without much thinking.

Changes the previous API when the color is not found: rather than
returning `false`, throw an Exception and properly notify the developer
of the color doing that. Using functions with multiple return types
causes you to use `if`s in the code without necessity.

The script provided to generate the classes from the JSON is added to
source control for reproducibility purposes, it can be safely removed on
next versions (as it is already under version control) as a TODO
states.

There are still some low hanging fruits declared on @TODO annotations in
the code. Others, not so clear, are:

* Depend on a `FileInfo` class rather than using file functions inside
  classes.
* Saving an image to a given path is a repeated code which can clearly
  be in a separate class.
* The ChangeColor feels like a [Factory][] but it is not one. Accepting
  an image on its constructor makes difficult to re-use it to do other
  operations. Maybe, the `ChangeColor` can create a `ColorChange`
  instance which accepts a `from` and `to` argument into a `save`
  method.

[Factory]: https://sourcemaking.com/design_patterns/factory_method